requirements.txt
pytest      6.2.2
PySide2     5.15.2
mss         6.1.0
keyboard    0.13.5
Pillow      8.2.0
numpy       1.20.2
onnxruntime 1.7.0
opencv-python-headless 4.5.1.48
Pinyin      0.4.0